[#7] enum 파일 분리
